+++ plugins/analysis-kuromoji/src/main/java/org/elasticsearch/plugin/analysis/kuromoji/AnalysisKuromojiPlugin.java
@@ -11,15 +11,6 @@ package org.elasticsearch.plugin.analysis.kuromoji;
 import org.apache.lucene.analysis.Analyzer;
 import org.elasticsearch.index.analysis.AnalyzerProvider;
 import org.elasticsearch.index.analysis.CharFilterFactory;
-import org.elasticsearch.index.analysis.JapaneseStopTokenFilterFactory;
-import org.elasticsearch.index.analysis.KuromojiAnalyzerProvider;
-import org.elasticsearch.index.analysis.KuromojiBaseFormFilterFactory;
-import org.elasticsearch.index.analysis.KuromojiIterationMarkCharFilterFactory;
-import org.elasticsearch.index.analysis.KuromojiKatakanaStemmerFactory;
-import org.elasticsearch.index.analysis.KuromojiNumberFilterFactory;
-import org.elasticsearch.index.analysis.KuromojiPartOfSpeechFilterFactory;
-import org.elasticsearch.index.analysis.KuromojiReadingFormFilterFactory;
-import org.elasticsearch.index.analysis.KuromojiTokenizerFactory;
 import org.elasticsearch.index.analysis.TokenFilterFactory;
 import org.elasticsearch.index.analysis.TokenizerFactory;
 import org.elasticsearch.indices.analysis.AnalysisModule.AnalysisProvider;
