+++ plugins/analysis-kuromoji/src/main/java/org/elasticsearch/plugin/analysis/kuromoji/KuromojiReadingFormFilterFactory.java
@@ -6,25 +6,26 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.plugin.analysis.kuromoji;
 
 import org.apache.lucene.analysis.TokenStream;
-import org.apache.lucene.analysis.ja.JapaneseKatakanaStemFilter;
+import org.apache.lucene.analysis.ja.JapaneseReadingFormFilter;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.IndexSettings;
+import org.elasticsearch.index.analysis.AbstractTokenFilterFactory;
 
-public class KuromojiKatakanaStemmerFactory extends AbstractTokenFilterFactory {
+public class KuromojiReadingFormFilterFactory extends AbstractTokenFilterFactory {
 
-    private final int minimumLength;
+    private final boolean useRomaji;
 
-    public KuromojiKatakanaStemmerFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
+    public KuromojiReadingFormFilterFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
         super(indexSettings, name, settings);
-        minimumLength = settings.getAsInt("minimum_length", JapaneseKatakanaStemFilter.DEFAULT_MINIMUM_LENGTH);
+        useRomaji = settings.getAsBoolean("use_romaji", false);
     }
 
     @Override
     public TokenStream create(TokenStream tokenStream) {
-        return new JapaneseKatakanaStemFilter(tokenStream, minimumLength);
+        return new JapaneseReadingFormFilter(tokenStream, useRomaji);
     }
 }
