+++ plugins/analysis-kuromoji/src/test/resources/org/elasticsearch/plugin/analysis/kuromoji/user_dict.txt
